+++ lib/src/search_app_bar_page/controller/searcher_page_pagination_future_controller.dart
@@ -297,7 +297,7 @@ class SearcherPagePaginationFutureController<T> extends SeacherBase {
   }
 }
 
-//@immutable
+@immutable
 class AsyncSnapshotScrollPage<T> {
   final AsyncSnapshot<List<T>> snapshot;