+++ test/data_fake_api/list_person.dart
@@ -1,0 +1,66 @@
+final dataListPerson3 = <Person>[
+  Person(name: 'Rafaela Pinho', age: 30),
+  Person(name: 'Paulo Emilio Silva', age: 45),
+  Person(name: 'Pedro Gomes', age: 18),
+  Person(name: 'Orlando Guerra', age: 23),
+  Person(name: 'Ana Pereira', age: 23),
+  Person(name: 'Zacarias Triste', age: 15),
+  Person(name: 'Antonio Rabelo', age: 33),
+  Person(name: 'Leticia Maciel', age: 47),
+  Person(name: 'Patricia Oliveira', age: 19),
+  Person(name: 'Pedro Lima', age: 15),
+  Person(name: 'Fabio Melo', age: 51),
+  Person(name: 'Junior Rabelo', age: 33),
+  Person(name: 'Lucia Maciel', age: 47),
+  Person(name: 'Ana Oliveira', age: 19),
+  Person(name: 'Thiago Silva', age: 33),
+  Person(name: 'Charles Ristow', age: 47),
+  Person(name: 'Raquel Montenegro', age: 19),
+  Person(name: 'Rafael Peireira', age: 15),
+  Person(name: 'Thiago Ferreira', age: 33),
+  Person(name: 'Joaquim Gomes', age: 18),
+  Person(name: 'Esther Guerra', age: 23),
+  Person(name: 'Pedro Braga', age: 19),
+  Person(name: 'Milu Silva', age: 17),
+  Person(name: 'William Ristow', age: 47),
+  Person(name: 'Elias Tato', age: 22),
+  Person(name: 'Dada Istomesmo', age: 44),
+  Person(name: 'Nome Incomum', age: 52),
+  Person(name: 'Qualquer Nome', age: 9),
+  Person(name: 'First Last', age: 11),
+  Person(name: 'Bom Dia', age: 23),
+  Person(name: 'Bem Mequiz', age: 13),
+  Person(name: 'Mal Mequer', age: 71),
+  Person(name: 'Quem Sabe', age: 35),
+  Person(name: 'Miriam Leitao', age: 33),
+  Person(name: 'Gabriel Mentiroso', age: 19),
+  Person(name: 'Caio Petro', age: 27),
+  Person(name: 'Tanto Nome', age: 66),
+  Person(name: 'Nao Diga', age: 33),
+  Person(name: 'Fique Queto', age: 11),
+  Person(name: 'Cicero Gome', age: 37),
+  Person(name: 'Carlos Gome', age: 48),
+  Person(name: 'Mae Querida', age: 45),
+  Person(name: 'Exausto Nome', age: 81),
+];
+
+class Person {
+  final String name;
+
+  final int age;
+
+  Person({this.name, this.age});
+
+  @override
+  String toString() {
+    return 'Person{name: $name, age: $age}';
+  }
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is Person && runtimeType == other.runtimeType && name == other.name;
+
+  @override
+  int get hashCode => name.hashCode;
+}